# tests/contract/api/openapi_snapshot.json
{
 "components": {
  "schemas": {
...
   "CreateOrganizationRequest": {
    "additionalProperties": false,
    "properties": {
     "organization_id": {
      "format": "uuid",
      "title": "Organization Id",
      "type": "string"
     }
    },
    "required": [
     "organization_id"
    ],
    "title": "CreateOrganizationRequest",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "OrganizationRecord": {
    "properties": {
     "id": {
      "format": "uuid",
      "title": "Id",
      "type": "string"
     },
     "status": {
      "title": "Status",
      "type": "string"
     }
    },
    "required": [
     "id",
     "status"
    ],
    "title": "OrganizationRecord",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "zhiwei",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/organizations": {
   "get": {
    "operationId": "list_organizations_api_v1_organizations_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/OrganizationRecord"
         },
         "title": "Response List Organizations Api V1 Organizations Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "List Organizations",
    "tags": [
     "organizations"
    ]
   },
   "post": {
    "operationId": "bootstrap_organization_api_v1_organizations_post",
    "parameters": [
     {
      "in": "header",
      "name": "Idempotency-Key",
      "required": true,
      "schema": {
       "minLength": 1,
       "pattern": "\\S+",
       "title": "Idempotency-Key",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CreateOrganizationRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {}
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Bootstrap Organization",
    "tags": [
     "organizations"
    ]
   }
  },
  "/api/v1/organizations/{organization_id}": {
   "get": {
    "operationId": "get_organization_api_v1_organizations__organization_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "organization_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Organization Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/OrganizationRecord"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Organization",
    "tags": [
     "organizations"
    ]
   }
  },
...
 }
}